Order import Command and Job Written. New import queue mechanism implemented

+++ src/database/migrations/2021_09_25_091822_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('title')->index();
             $table->text('description')->nullable();
             $table->string('image')->nullable();
-            $table->integer('sku')->index();
+            $table->bigInteger('sku')->index();
             $table->float('price');
             $table->timestamps();
         });
